Fix build after 6b85c39b545d7568322c450a2d4d90f4df7994e6
Because of missing imports and some undefined variables the latest code wouldn't compile.

I have fixed it.

The only thing I am not sure about is dispatchConfig: PropTypes.bool.isRequired
Not sure if it's truly bool or if this is valid, but it works finally again, but with proper log timezone!